requirements.txt
fastapi
uvicorn
openai
chromadb
streamlit
python-dotenv
pandas
tqdm
sentence_transformers
groq
